fix model/framework small issues
