+++ libs/core/src/ports/interfaces.ts
@@ -1,4 +1,4 @@
-import { Events } from '@hicommonwealth/schemas';
+import { Events, OutboxEvents } from '@hicommonwealth/schemas';
 import { Readable } from 'stream';
 import { z } from 'zod';
 import {
@@ -466,7 +466,9 @@ export type RoutingKey =
   | Concat<EventNamesType, RoutingKeyTagsType>;
 
 export interface Broker extends Disposable {
-  publish<Name extends Events>(event: EventContext<Name>): Promise<boolean>;
+  publish<Name extends OutboxEvents>(
+    event: EventContext<Name>,
+  ): Promise<boolean>;
 
   subscribe<Inputs extends EventSchemas>(
     consumer: () => EventsHandlerMetadata<Inputs>,
