+++ libs/schemas/src/commands/token.schemas.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import { AuthContext } from '../context';
-import { LaunchpadTrade } from '../entities';
-import { TokenView } from '../queries';
+import { events } from '../events';
+import { LaunchpadTradeView, TokenView } from '../queries';
 
 export const CreateToken = {
   input: z.object({
@@ -18,11 +18,6 @@ export const CreateToken = {
   context: AuthContext,
 };
 
-export const LaunchpadTradeView = LaunchpadTrade.extend({
-  community_token_amount: z.string(),
-  floating_supply: z.string(),
-});
-
 export const CreateLaunchpadTrade = {
   input: z.object({
     eth_chain_id: z.number(),
@@ -31,51 +26,9 @@ export const CreateLaunchpadTrade = {
   output: LaunchpadTradeView.nullish(),
 };
 
-export const GetLaunchpadTrades = {
-  input: z.object({
-    token_address: z.string().optional(),
-    trader_addresses: z.string().optional(),
-  }),
-  output: LaunchpadTrade.extend({
-    community_token_amount: z.string(),
-    floating_supply: z.string(),
-    name: z.string(),
-    symbol: z.string(),
-    community_id: z.string(),
-    community_icon_url: z.string(),
-    user_id: z.number().nullish(),
-    user_name: z.string().nullish(),
-    user_avatar_url: z.string().nullish(),
-  }).array(),
-};
-
-export const GetTokenizedThreadsAllowed = {
-  input: z.object({
-    community_id: z.string(),
-    topic_id: z.number(),
-  }),
+export const ProjectLaunchpadTrade = {
+  input: events.LaunchpadTokenTraded,
   output: z.object({
-    tokenized_threads_enabled: z.boolean(),
-    thread_purchase_token: z.string().nullish(),
-  }),
-};
-
-export const GetTokenInfoAlchemy = {
-  input: z.object({
-    eth_chain_id: z.number(),
-    token_address: z.string(),
-  }),
-  output: z.object({
-    network: z.string(),
-    address: z.string(),
-    currency: z.string(),
-    data: z
-      .object({
-        value: z.string(),
-        timestamp: z.string(),
-        marketCap: z.string(),
-        totalVolume: z.string(),
-      })
-      .array(),
+    community_id: z.string().optional(),
   }),
 };
