+++ libs/schemas/src/queries/index.ts
@@ -6,6 +6,7 @@ export * from './eventStream.schemas';
 export * from './pagination';
 export * from './poll.schemas';
 export * from './quest.schemas';
+export * from './search.schemas';
 export * from './subscription.schemas';
 export * from './super-admin.schemas';
 export * from './tag.schemas';
