+++ libs/schemas/src/queries/token.schemas.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { LaunchpadToken } from '../entities';
+import { LaunchpadToken, LaunchpadTrade } from '../entities';
 import { PaginatedResultSchema, PaginationParamsSchema } from './pagination';
 
 export const TokenView = LaunchpadToken.extend({
@@ -34,4 +34,58 @@ export const GetToken = {
       },
     ),
   output: z.union([TokenView, z.null()]),
+};
+
+export const LaunchpadTradeView = LaunchpadTrade.extend({
+  community_token_amount: z.string(),
+  floating_supply: z.string(),
+});
+
+export const GetLaunchpadTrades = {
+  input: z.object({
+    token_address: z.string().optional(),
+    trader_addresses: z.string().optional(),
+  }),
+  output: LaunchpadTrade.extend({
+    community_token_amount: z.string(),
+    floating_supply: z.string(),
+    name: z.string(),
+    symbol: z.string(),
+    community_id: z.string(),
+    community_icon_url: z.string(),
+    user_id: z.number().nullish(),
+    user_name: z.string().nullish(),
+    user_avatar_url: z.string().nullish(),
+  }).array(),
+};
+
+export const GetTokenizedThreadsAllowed = {
+  input: z.object({
+    community_id: z.string(),
+    topic_id: z.number(),
+  }),
+  output: z.object({
+    tokenized_threads_enabled: z.boolean(),
+    thread_purchase_token: z.string().nullish(),
+  }),
+};
+
+export const GetTokenInfoAlchemy = {
+  input: z.object({
+    eth_chain_id: z.number(),
+    token_address: z.string(),
+  }),
+  output: z.object({
+    network: z.string(),
+    address: z.string(),
+    currency: z.string(),
+    data: z
+      .object({
+        value: z.string(),
+        timestamp: z.string(),
+        marketCap: z.string(),
+        totalVolume: z.string(),
+      })
+      .array(),
+  }),
 };
